requirements.txt
matplotlib==3.5.2
numpy==1.21.6
pandas==1.3.0
PyYAML==6.0
scikit_learn==1.0.2
seaborn==0.12.2
torch==1.12.1
torchvision==0.13.1
